move tests into their own file
